refactor(app): extract depth, audio-reactive, and share hooks (#913)
Move transformers depth pipeline, audio param modulation, and chain/hash
sharing into dedicated hooks wired back into App.tsx with unchanged behavior.
