+++ src/hooks/index.ts
@@ -9,3 +9,12 @@ export type { UseStorageReturn } from './useStorage';
 export { useWASM } from './useWASM';
 export { useAudioAnalyzer } from './useAudioAnalyzer';
 export { usePerformanceMonitor } from './usePerformanceMonitor';
+
+export { useDepthEstimation } from './useDepthEstimation';
+export type { UseDepthEstimationOptions, UseDepthEstimationReturn } from './useDepthEstimation';
+
+export { useAudioReactiveParams } from './useAudioReactiveParams';
+export type { UseAudioReactiveParamsOptions, UseAudioReactiveParamsReturn } from './useAudioReactiveParams';
+
+export { useShareChain } from './useShareChain';
+export type { UseShareChainOptions, UseShareChainReturn } from './useShareChain';
